Binder de numero flexivel: aceita virgula ou ponto em decimais
NumeroFlexivelModelBinder faz o binding de double/float/decimal aceitar
tanto "10,5" quanto "10.5", independente da cultura do servidor
(pt-BR quebrava coordenadas de escalacao). EscalacaoInput.PosicaoX/Y
passam de int para double para nao truncar as casas decimais.

+++ ControleFutebolWeb/Models/EscalacaoInput.cs
@@ -5,8 +5,10 @@
         public int Id { get; set; }  // ← adicionar
         public int JogadorId { get; set; }
         public int PosicaoId { get; set; }
-        public int PosicaoX { get; set; }
-        public int PosicaoY { get; set; }
+        // double: as coordenadas dos slots de formação aceitam casas decimais
+        // (Escalacao.PosicaoX/Y também são double) — int truncava ao salvar.
+        public double PosicaoX { get; set; }
+        public double PosicaoY { get; set; }
     }
 
 }
